Continued implementing Multiplayer Setup Screen.
